fix: auth login #23

+++ Project/frontend/src/context/authContext.tsx
@@ -1,10 +1,12 @@
 import { createContext, useContext, useState } from "react";
+import Swal from "sweetalert2";
 
 interface Usuario {
     token: string
     user: {
         EDV: number
         tipo: string
+        ativo:boolean
     }
 }
 
@@ -22,7 +24,19 @@ export function AuthProvider({children}: {children: React.ReactNode}) {
         return salvo ? JSON.parse(salvo) : null;
     });
     function login(usuario: Usuario){
+        if (!usuario.user.ativo) {
+            Swal.fire({
+                icon: "error",
+                title: "Usuário inativo",
+                text: "Sua conta está desativada.",
+                confirmButtonColor: "#2B83D5",
+                confirmButtonText: "OK"
+            });
+    
+            return;
+        }
         setUsuario(usuario);
+        console.log("User "+ usuario.user)
         localStorage.setItem(
             "usuario",
             JSON.stringify(usuario.user)
@@ -31,6 +45,7 @@ export function AuthProvider({children}: {children: React.ReactNode}) {
             "token",
             JSON.stringify(usuario.token)
         );
+       
     }
     function logout(){
         setUsuario(null);
